+++ src/components/plips-layout.js
@@ -20,16 +20,11 @@ import {
   NATIONWIDE_SCOPE,
   STATEWIDE_SCOPE,
   CITYWIDE_SCOPE,
-  findByStrIndex,
 } from "../utils";
 
 import Icon from "react-native-vector-icons/MaterialIcons";
 
 import * as Animatable from "react-native-animatable";
-import {
-  Parallax,
-  ScrollDriver,
-} from "@shoutem/animation";
 
 import Layout from "./layout";
 import NavigationBar from "./navigation-bar";
@@ -38,6 +33,7 @@ import HeaderLogo from "./header-logo";
 import NetworkImage from "./network-image";
 import LinearGradient from "react-native-linear-gradient";
 import FlatButton from "./flat-button";
+import TransparentFlatButton from "./transparent-flat-button";
 import Collapsable from "./collapsable";
 import MyListView from "./list-view";
 import Triangle from "./triangle";
@@ -47,14 +43,8 @@ import styles from "../styles/plips-layout";
 
 import locale from "../locales/pt-BR";
 
-const LINKS_KEY = "LINKS";
-
 
 export default class PlipsLayout extends Component {
-  nationwideDriver = new ScrollDriver();
-  statewideDriver = new ScrollDriver();
-  citywideDriver = new ScrollDriver();
-
   state = {
     shouldFetchFirstTimeStatePlips: true,
     shouldFetchFirstTimeCityPlips: true,
@@ -88,6 +78,7 @@ export default class PlipsLayout extends Component {
     openMenu: PropTypes.func.isRequired,
     plipsSignInfo: PropTypes.object.isRequired,
     statewidePlipsDataSource: PropTypes.instanceOf(ListView.DataSource).isRequired,
+    userSignInfo: PropTypes.object.isRequired,
     onChangeScope: PropTypes.func.isRequired,
     onFetchPlips: PropTypes.func.isRequired,
     onFetchPlipsNextPage: PropTypes.func.isRequired,
@@ -191,7 +182,6 @@ export default class PlipsLayout extends Component {
 
     const onScroll = event => {
       this.nationwideScroll = event.nativeEvent.contentOffset;
-      this.nationwideDriver.scrollViewProps.onScroll(event);
     };
 
     const scrollTo = this.nationwideScroll;
@@ -203,7 +193,6 @@ export default class PlipsLayout extends Component {
       dataSource,
       onScroll,
       scrollTo,
-      driver: this.nationwideDriver,
     });
   }
 
@@ -219,7 +208,6 @@ export default class PlipsLayout extends Component {
 
     const onScroll = event => {
       this.statewideScroll = event.nativeEvent.contentOffset;
-      this.statewideDriver.scrollViewProps.onScroll(event);
     };
     const scrollTo = this.statewideScroll;
 
@@ -230,7 +218,6 @@ export default class PlipsLayout extends Component {
       dataSource,
       onScroll,
       scrollTo,
-      driver: this.statewideDriver,
     });
   }
 
@@ -246,7 +233,6 @@ export default class PlipsLayout extends Component {
 
     const onScroll = event => {
       this.citywideScroll = event.nativeEvent.contentOffset;
-      this.citywideDriver.scrollViewProps.onScroll(event);
     };
     const scrollTo = this.citywideScroll;
 
@@ -257,14 +243,12 @@ export default class PlipsLayout extends Component {
       dataSource,
       onScroll,
       scrollTo,
-      driver: this.citywideDriver,
     });
   }
 
-  renderScopeContent({ dataSource, error, isFetching, isRefreshing, onScroll, scrollTo, driver }) {
+  renderScopeContent({ dataSource, error, isFetching, isRefreshing, onScroll, scrollTo }) {
     const { filters } = this.props;
     const hasRows = dataSource.getRowCount() > 0;
-    const hasNoListYet = error || isFetching || !hasRows;
     const shouldShowNoPlips =
       !error &&
       !isFetching &&
@@ -277,7 +261,7 @@ export default class PlipsLayout extends Component {
 
 
     return (
-      <View style={{flex: 1, backgroundColor: hasNoListYet ? "white" : "black" }}>
+      <View style={{flex: 1, backgroundColor: "#F9F9F9" }}>
         {
           !error && hasRows &&
             this.renderListView({
@@ -285,7 +269,6 @@ export default class PlipsLayout extends Component {
               isRefreshing,
               onScroll,
               scrollTo,
-              driver,
             })
         }
 
@@ -308,7 +291,7 @@ export default class PlipsLayout extends Component {
     );
   }
 
-  renderListView({ dataSource, isRefreshing, onScroll, scrollTo, driver }) {
+  renderListView({ dataSource, isRefreshing, onScroll, scrollTo }) {
     const {
       onFetchPlipsNextPage,
       onRefresh,
@@ -316,16 +299,16 @@ export default class PlipsLayout extends Component {
 
     return (
       <MyListView
-        {...driver.scrollViewProps}
-
         scrollTo={scrollTo}
         style={styles.listView}
+        contentContainerStyle={styles.listViewContent}
         automaticallyAdjustContentInsets={false}
         enableEmptySections={true}
         onEndReached={onFetchPlipsNextPage}
         onEndReachedThreshold={300}
+        scrollEventThrottle={500}
         dataSource={dataSource}
-        renderRow={this.renderRow({ height: 333, margin: 0, dataSource, driver })}
+        renderRow={this.renderRow({ height: 360, margin: 0 })}
         onScroll={onScroll}
         refreshControl={
           <RefreshControl
@@ -338,136 +321,77 @@ export default class PlipsLayout extends Component {
     );
   }
 
-  renderRow = ({ height, margin, dataSource, driver }) => ([plip, index], section, row, highlightRow) => {
+  renderRow = ({ height, margin }) => ([plip], section, row, highlightRow) => {
     const { onGoToPlip } = this.props;
-
-    const shouldHideOverflow = index > 0 || dataSource.getRowCount() === 1;
-
-    const isLink = plip.key === LINKS_KEY;
-    const TouchableView = isLink ? View : TouchableOpacity;
 
     return (
       <View style={styles.rowContainer}>
-        <TouchableView
+        <TouchableOpacity
           onPress={() => {
             highlightRow(section, row);
             onGoToPlip(plip);
           }}
           style={[styles.tableRow, {
-            // We don't want to overflow the first row, so we can see the image
-            // on ios during the bounce animation
-            overflow: shouldHideOverflow ? "hidden" : "visible",
-
-            height,
+            minHeight: height,
             margin,
           }]}
         >
-          {isLink && this.renderRowLinks()}
-          {!isLink && this.renderRowPlip({ plip, index, height, margin, driver })}
-        </TouchableView>
+          {this.renderRowPlip({ plip, height, margin })}
+        </TouchableOpacity>
       </View>
     );
   }
 
-  renderRowLinks() {
-    const { mudamos, projectsReason } = siteLinks.homeLinks;
-    const { onOpenURL } = this.props;
+  renderRowPlip({ plip }) {
+    const { plipsSignInfo, userSignInfo, onGoToPlip } = this.props;
+    const plipSignInfo = plipsSignInfo[plip.id];
+    const plipUserSignInfo = userSignInfo[plip.id];
+    const hasSigned = !!(plipUserSignInfo && plipUserSignInfo.updatedAt);
+
+    // Not every animation seem to work on both platforms
+    const AnimatableView = Platform.OS === "ios" ? Animatable.View : View;
 
     return (
-      <LinearGradient
-        start={[0.0, 0.1]}
-        end={[0.5, 1.0]}
-        locations={[0, 0.5]}
-        style={styles.footerContainer}
-        colors={["#9844ce", "#7E52D8"]}
-      >
-        <TouchableOpacity
-          onPress={() => onOpenURL(mudamos.link) }
-          style={styles.full}
-        >
-          {this.renderLink({ ...mudamos, icon: "extension" })}
-        </TouchableOpacity>
-
-        <View style={styles.hairline} />
-
-        <TouchableOpacity
-          onPress={() => onOpenURL(projectsReason.link) }
-          style={styles.full}
-        >
-          {this.renderLink({ ...projectsReason, icon: "info" })}
-        </TouchableOpacity>
-
-      </LinearGradient>
-    );
-  }
-
-  renderLink({ title, icon }) {
-    return (
-      <View style={styles.actionRow}>
-        <Icon
-          name={icon}
-          size={40}
-          color="#fff"
-          style={styles.actionIcon}
+      <View style={styles.plipRow}>
+        <NetworkImage
+          source={{uri: this.plipImage(plip)}}
+          resizeMode="cover"
+          style={styles.plipImage}
         />
-        <Text style={styles.actionTitle}>{title.toUpperCase()}</Text>
-        <Icon
-          name="chevron-right"
-          size={40}
-          color="#fff"
-        />
-      </View>
-    );
-  }
-
-  renderRowPlip({ plip, index, height, margin, driver }) {
-    const { plipsSignInfo } = this.props;
-    const plipSignInfo = findByStrIndex(plip.id, plipsSignInfo);
-
-    const totalHeight = height + margin;
-    const scrollRange = totalHeight * (index - 1);
-
-    // For some reason, Android crashes and a transform is applied
-    // TODO: upgrade react-native and test again
-    const ParallaxView = Platform.OS == "ios" ? Parallax : View;
-
-    return (
-      <View style={styles.full}>
-        <ParallaxView
-          driver={driver}
-          scrollSpeed={0.7}
-          style={{alignItems: "center", justifyContent: "center"}}
-          header={index == 0 /* If this is not informed, Shouten will use the incorrect initial transform */}
-          extrapolation={{
-            // We must "divide" the scroll range into each PLIP,
-            // otherwise the accumulated error will cause layout break on long lists
-            inputRange: [scrollRange, scrollRange + totalHeight],
-          }}
-        >
-          <NetworkImage
-            source={{uri: this.plipImage(plip)}}
-            resizeMode="cover"
-            style={styles.plipImage}
-          />
-        </ParallaxView>
 
         { /* This gradient improves the reading of the PLIP title and subtitle */ }
         <LinearGradient
-          colors={["rgba(0, 0, 0, 0)", "rgba(0, 0, 0, 0.8)", "rgba(0, 0, 0, 1)"]}
-          locations={[0.3, 0.7, 1]}
+          colors={["rgba(0, 0, 0, .3)", "rgba(0, 0, 0, 0.3)", "rgba(0, 0, 0, .7)"]}
+          locations={[0, 0.7, 1]}
           style={styles.plipImageGradient}
         />
 
         <View style={styles.plipTitleContainer}>
           <View style={styles.plipTitleInnerContainer}>
-            <Text style={styles.plipTitle} numberOfLines={3}>
+            <Text style={styles.plipTitle}>
               {plip.phase.name}
             </Text>
 
-            <Text style={styles.plipSubtitle} numberOfLines={3}>
+            <Text style={styles.plipSubtitle}>
               {plip.phase.description}
             </Text>
           </View>
+
+          <TransparentFlatButton
+            title={locale.moreDetails.toUpperCase()}
+            onPress={() => onGoToPlip(plip)}
+            style={{
+              height: 30,
+              marginHorizontal: 20,
+              marginTop: 55,
+              marginBottom: 25,
+            }}
+            textStyle={{
+              textShadowColor: "rgba(0,0,0, 1)",
+              textShadowOffset: { width: 1, height: 1 },
+              textShadowRadius: 1,
+            }}
+          />
 
           {
             !!plipSignInfo &&
@@ -479,6 +403,23 @@ export default class PlipsLayout extends Component {
                 />
               </Animatable.View>
           }
+
+          {
+            hasSigned &&
+              <AnimatableView animation="zoomIn" easing="ease-in-out-sine" style={styles.signedContainer} duration={2000}>
+                <LinearGradient
+                  colors={["#00DB5E", "#00A79E"]}
+                  style={styles.signedGradient}
+                >
+                  <Text style={styles.signedText}>{locale.signed}</Text>
+                </LinearGradient>
+              </AnimatableView>
+          }
+
+          <Image source={require("../images/plips-top-left.png")} style={{position: "absolute", top: 0, left: 0}} />
+          <Image source={require("../images/plips-bottom-right.png")} style={{position: "absolute", bottom: 0, right: 0}} />
+          <Image source={require("../images/plips-bottom-left.png")} style={{position: "absolute", bottom: 0, left: 0}} />
+          <Image source={require("../images/plips-top-right.png")} style={{position: "absolute", top: 0, right: 0}} />
         </View>
 
       </View>
@@ -593,7 +534,7 @@ export default class PlipsLayout extends Component {
         <Icon
           name="arrow-drop-down"
           size={20}
-          color="rgba(255,255,255,0.3)"
+          color="#8934E5"
         />
       </TouchableOpacity>
     );
@@ -673,11 +614,11 @@ export default class PlipsLayout extends Component {
 }
 
 const TabItem = ({ selected, title, onPress }) => {
-  const opacity = selected ? 1 : 0.43;
+  const color = selected ? "#8934E5"  : "#AAAAAA";
 
   return (
     <TouchableOpacity onPress={onPress}>
-      <Text style={[styles.tabItem, { color: `rgba(255, 255, 255, ${opacity})` }]}>
+      <Text style={[styles.tabItem, { color }]}>
         {title}
       </Text>
     </TouchableOpacity>